requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.33
attrs>=19.1.0
absl-py>=0.10.0
flatbuffers>=2.0
scikit-learn==1.4.2
scipy>=1.6.0
threadpoolctl>=2.0.0
numpy==1.26.4
pandas==2.2.2
joblib==1.3.2
python-telegram-bot==20.8
Pillow==10.3.0
matplotlib==3.8.4
seaborn==0.13.2
